# Conflicts:
#	Handin3/Code/AirTrafficMonitor.Unit.Test/AirTrafficMonitor.Unit.Test.csproj
#	Handin3/Code/AirTrafficMonitor.Unit.Test/Class/FlightTrackerSingleTest.cs

+++ Handin3/Code/AirTrafficMonitor/Classes/FlightCalculator/FlightVelocityCalculator.cs
@@ -10,23 +10,23 @@ namespace AirTrafficMonitor
 {
     public class FlightVelocityCalculator : IFlightVelocityCalculator
     {
-        public double CalculateCurrentVelocity(Vector2 previous_position, DateTime? previous_timestamp, Vector2 current_position, DateTime current_timestamp)
+        public float CalculateCurrentVelocity(Vector2 previous_position, DateTime? previous_timestamp, Vector2 current_position, DateTime current_timestamp)
         { 
 
-            double distance, deltaY, deltaX;
+            float distance, deltaY, deltaX;
             if (current_position.X > previous_position.X) deltaX = current_position.X - previous_position.X;
             else deltaX = previous_position.X - current_position.X;
             if (current_position.Y > previous_position.Y) deltaY = current_position.Y - previous_position.Y;
             else deltaY = previous_position.Y - current_position.Y;
-            distance = Math.Sqrt(Math.Pow(deltaX, 2) + Math.Pow(deltaY, 2));
+            distance = (float)Math.Sqrt(Math.Pow(deltaX, 2) + Math.Pow(deltaY, 2));
 
-            double timeBetweenInSeconds;
+            float timeBetweenInSeconds;
             var DatetimeToTimeStamp = new DateTime();
             TimeSpan current, previous, resultTimespan;
             current = DatetimeToTimeStamp.Subtract(current_timestamp);
             previous = DatetimeToTimeStamp.Subtract(previous_timestamp ?? DateTime.Now);
             resultTimespan = current - previous;
-            timeBetweenInSeconds = (double)resultTimespan.TotalSeconds;
+            timeBetweenInSeconds = (float)resultTimespan.TotalSeconds;
             if (timeBetweenInSeconds < 0)
             {
                 timeBetweenInSeconds = timeBetweenInSeconds * -1; //convert to positive
